python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "av>=11.0.0",
    "google-cloud-speech>=2.28.1",
    "google-crc32c>=1.6.0",
    "huddle01>=1.0.0",
    "numpy>=2.1.3",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.1",
    "sounddevice>=0.5.1",
    "soundfile>=0.12.1",
]
